+++ TelemetryWebApp/frontend-web-telemetry/src/components/plots/VehicleSpeed.vue
@@ -7,6 +7,7 @@
 <script setup lang="ts">
 import {computed, onMounted, onUnmounted, watch} from "vue";
 import { useFileStoreStore } from "@/stores/FileStore";
+import { chartSyncService } from "@/services/chartSync";
 import {
   SciChartSurface,
   NumericAxis,
@@ -15,7 +16,9 @@ import {
   XyDataSeries,
   SciChartJSDarkTheme,
   ZoomPanModifier,
-  ZoomExtentsModifier
+  ZoomExtentsModifier,
+  MouseWheelZoomModifier,
+  RolloverModifier
 } from "scichart";
 
 // Retrieve data from store
@@ -64,10 +67,12 @@ onMounted(async () => {
     dataSeries: speedDataSeries,
   }));
 
-    // Modifiers
+  const modifierGroup = chartSyncService.modifierGroupId;
   sciChartSurface.chartModifiers.add(
-      new ZoomPanModifier(),
-      new ZoomExtentsModifier()
+    new ZoomPanModifier({ modifierGroup }),
+    new MouseWheelZoomModifier({ modifierGroup }),
+    new ZoomExtentsModifier({ modifierGroup }),
+    new RolloverModifier({ modifierGroup })
   );
 
    // Axes
@@ -83,8 +88,12 @@ onMounted(async () => {
 
 });
 
-onUnmounted(async () => {
-  sciChartSurface.delete();
+onUnmounted(() => {
+  if (sciChartSurface) {
+    chartSyncService.unregister(sciChartSurface);
+    sciChartSurface.delete();
+    sciChartSurface = null;
+  }
 });
 
 </script>
@@ -96,3 +105,4 @@ onUnmounted(async () => {
   height: 240px;
 }
 </style>
+  chartSyncService.register(sciChartSurface);
