+++ TelemetryWebApp/frontend-web-telemetry/src/components/plots/Voltage.vue
@@ -7,6 +7,7 @@
 <script setup lang="ts">
 import {computed, watch, onMounted, onUnmounted} from "vue";
 import { useFileStoreStore } from "@/stores/FileStore";
+import { chartSyncService } from "@/services/chartSync";
 import {
   SciChartSurface,
   NumericAxis,
@@ -15,7 +16,9 @@ import {
   XyDataSeries,
   SciChartJSDarkTheme,
   ZoomPanModifier,
-  ZoomExtentsModifier
+  ZoomExtentsModifier,
+  MouseWheelZoomModifier,
+  RolloverModifier
 } from "scichart";
 
 // Retrieve data from store
@@ -58,7 +61,16 @@ onMounted(async () => {
 
   sciChartSurface.xAxes.add(new NumericAxis(wasmContext, { autoRange: EAutoRange.Always, drawLabels: false}));
   sciChartSurface.yAxes.add(new NumericAxis(wasmContext, { autoRange: EAutoRange.Always}));
-  sciChartSurface.chartModifiers.add(new ZoomPanModifier(), new ZoomExtentsModifier());
+
+  const modifierGroup = chartSyncService.modifierGroupId;
+  sciChartSurface.chartModifiers.add(
+    new ZoomPanModifier({ modifierGroup }),
+    new MouseWheelZoomModifier({ modifierGroup }),
+    new ZoomExtentsModifier({ modifierGroup }),
+    new RolloverModifier({ modifierGroup })
+  );
+
+  chartSyncService.register(sciChartSurface);
 
   updateChart();
 
@@ -67,8 +79,12 @@ onMounted(async () => {
   })
 });
 
-onUnmounted(async () => {
-  sciChartSurface.delete();
+onUnmounted(() => {
+  if (sciChartSurface) {
+    chartSyncService.unregister(sciChartSurface);
+    sciChartSurface.delete();
+    sciChartSurface = null;
+  }
 });
 
 </script>
